Make Todo Page component
TodoPage 컴포넌트를 만들어서 관심사를 분리.
